Refactor settings configuration
