+++ backend/core/boot_guard.py
@@ -1,8 +1,9 @@
-from backend.core.settings import settings
+from backend.core.settings import Settings, get_settings
 import re
 
 def assert_safe_database_url():
-    if settings.ENV == "prod":
+    settings = get_settings()
+    if settings.env == "prod":
         # Example: enforce hostname or db name pattern
         if settings.DATABASE_NAME_EXPECTED and not re.search(
             rf"/{re.escape(settings.DATABASE_NAME_EXPECTED)}(\?|$)",
