+++ backend/core/settings.py
@@ -3,10 +3,15 @@ from typing_extensions import  Optional, List
 from pydantic import Field
 from pydantic_settings import BaseSettings, SettingsConfigDict
 from functools import lru_cache
+import os
+
+def env_file_path() -> str:
+    env = os.getenv("ENV", "dev")
+    return f"backend/.env.{env}"
 
 class Settings(BaseSettings):
     model_config = SettingsConfigDict(
-        env_file="backend/.env",
+        env_file=env_file_path(),
         env_file_encoding="utf-8",
         extra="ignore",
         case_sensitive=False,
@@ -14,19 +19,22 @@ class Settings(BaseSettings):
 
     # App env
     env: str = Field(default="dev")  # dev|staging|prod
-
+    DATABASE_URL: str
+    ALLOW_DESTRUCTIVE_ENDPOINTS: bool = False
     # Security / JWT
-    jwt_secret_key: str
+    jwt_secret_key: str = Field(alias="JWT_SECRET_KEY")
     jwt_algorithm: str = "HS256"
     access_token_expiry: int = 30
     encrypt_algorithm: str = "HS256"
-    pgp_secret_key: str
+    pgp_secret_key: str  = Field(alias="PGP_SECRET_KEY")
 
+    ''' NO NEED BECAUSE url is provided directly
     # Postgres (runtime)
     postgres_host: str
     postgres_db: str
     postgres_user: str
     postgres_password: str
+    '''
 
     # eBay
     ebay_app_id: str | None = None
